+++ index.tsx
@@ -2,7 +2,6 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './App';
 import { AppProvider } from './store/AppContext';
-import './index.css';
 
 const rootElement = document.getElementById('root');
 if (!rootElement) {